refactor(cli): move bundle to cli and remove bundler
