+++ package.json
@@ -44,7 +44,6 @@
     "@babel/preset-env": "^7.10.0",
     "@babel/preset-react": "^7.10.0",
     "@babel/runtime": "^7.10.0",
-    "@carbon/bundler": "workspace:packages/bundler",
     "@carbon/cli": "workspace:packages/cli",
     "@commitlint/cli": "^8.3.5",
     "@commitlint/config-conventional": "^8.3.4",
@@ -146,7 +145,7 @@
     "preset": "jest-config-carbon",
     "collectCoverageFrom": [
       "packages/**/src/**/*.js",
-      "!packages/{bundler,cli,components,sketch}/**",
+      "!packages/{cli,components,sketch}/**",
       "!packages/**/{examples,stories}/**",
       "!**/*-story.js"
     ],
