--- packages/bundler/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "rules": {
-    "no-console": "OFF"
-  }
-}
