--- packages/bundler/.npmignore
@@ -1,4 +1,0 @@
-**/__mocks__/**
-**/__tests__/**
-**/examples/**
-**/tasks/**
