--- packages/bundler/README.md
@@ -1,29 +1,0 @@
-# @carbon/bundler
-
-> Bundler used for libraries in the Carbon Design System
-
-## Getting started
-
-To install `@carbon/bundler` in your project, you will need to run the following
-command using [npm](https://www.npmjs.com/):
-
-```bash
-npm install -S @carbon/bundler
-```
-
-If you prefer [Yarn](https://yarnpkg.com/en/), use the following command
-instead:
-
-```bash
-yarn add @carbon/bundler
-```
-
-## 🙌 Contributing
-
-We're always looking for contributors to help us fix bugs, build new features,
-or help us improve the project documentation. If you're interested, definitely
-check out our [Contributing Guide](/.github/CONTRIBUTING.md)! 👀
-
-## 📝 License
-
-Licensed under the [Apache 2.0 License](/LICENSE).
