--- packages/bundler/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "@carbon/bundler",
-  "description": "Bundler used for libraries in the Carbon Design System",
-  "version": "10.11.0",
-  "license": "Apache-2.0",
-  "bin": "./bin/bundler.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/carbon-design-system/carbon.git",
-    "directory": "packages/bundler"
-  },
-  "bugs": "https://github.com/carbon-design-system/carbon/issues",
-  "keywords": [
-    "ibm",
-    "elements",
-    "carbon",
-    "carbon-elements",
-    "carbon-design-system",
-    "components",
-    "react"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@babel/core": "^7.10.0",
-    "@babel/preset-env": "^7.10.0",
-    "@carbon/cli-reporter": "^10.4.0",
-    "@commitlint/parse": "^8.3.4",
-    "autoprefixer": "^9.7.4",
-    "babel-plugin-macros": "^2.4.2",
-    "chalk": "^2.4.1",
-    "change-case": "^4.1.1",
-    "cli-table": "^0.3.1",
-    "commander": "^2.19.0",
-    "cssnano": "^4.1.7",
-    "execa": "^2.0.3",
-    "filesize": "^3.6.1",
-    "fs-extra": "^8.1.0",
-    "glob": "^7.1.3",
-    "gzip-size": "^5.0.0",
-    "inquirer": "^6.5.0",
-    "klaw-sync": "^6.0.0",
-    "markdown-toc": "^1.2.0",
-    "node-sass": "^4.13.1",
-    "postcss": "^7.0.26",
-    "postcss-discard-comments": "^4.0.1",
-    "replace-in-file": "^3.4.4",
-    "rollup": "^0.67.1",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-commonjs": "^9.2.0",
-    "rollup-plugin-node-resolve": "^3.4.0",
-    "sassdoc": "^2.5.1",
-    "semver": "^6.3.0"
-  }
-}
