--- packages/bundler/src/tools/findPackageFolder.js
@@ -1,28 +1,0 @@
-/**
- * Copyright IBM Corp. 2018, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-'use strict';
-
-const fs = require('fs-extra');
-const path = require('path');
-
-async function findPackageFolder(entrypoint) {
-  let packageFolder = entrypoint;
-
-  while (packageFolder !== '/' && path.dirname(packageFolder) !== '/') {
-    packageFolder = path.dirname(packageFolder);
-    const packageJsonPath = path.join(packageFolder, 'package.json');
-
-    if (await fs.pathExists(packageJsonPath)) {
-      break;
-    }
-  }
-
-  return packageFolder;
-}
-
-module.exports = findPackageFolder;
