+++ packages/cli/package.json
@@ -27,7 +27,11 @@
     "@octokit/plugin-retry": "^2.2.0",
     "@octokit/plugin-throttling": "^2.6.0",
     "@octokit/rest": "^16.28.1",
+    "@rollup/plugin-babel": "^5.2.2",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.1",
     "chalk": "^2.4.2",
+    "change-case": "^4.1.2",
     "child-process-promise": "^2.2.1",
     "clipboardy": "^2.1.0",
     "enquirer": "^2.3.6",
@@ -42,6 +46,7 @@
     "progress-estimator": "^0.2.2",
     "remark": "^10.0.1",
     "replace-in-file": "^6.1.0",
+    "rollup": "^2.36.1",
     "sass": "^1.32.4",
     "sassdoc": "^2.7.3",
     "yargs": "^15.1.0"
