+++ packages/cli/src/commands/bundle/bundlers.js
@@ -8,11 +8,7 @@
 'use strict';
 
 const javascript = require('./javascript');
-const scss = require('./scss');
 
-const bundlers = new Map([
-  ['.js', javascript],
-  ['.scss', scss],
-]);
+const bundlers = new Map([['.js', javascript]]);
 
 module.exports = bundlers;
