+++ packages/colors/package.json
@@ -31,11 +31,10 @@
     "access": "public"
   },
   "scripts": {
-    "build": "yarn clean && bundler bundle src/index.js --name CarbonColors && node tasks/build.js && carbon-cli check \"scss/*.scss\" && carbon-cli sassdoc \"scss/*.scss\"",
+    "build": "yarn clean && carbon-cli bundle src/index.js --name CarbonColors && node tasks/build.js && carbon-cli check \"scss/*.scss\" && carbon-cli sassdoc \"scss/*.scss\"",
     "clean": "rimraf css es lib umd scss"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli": "^10.17.0",
     "@carbon/cli-reporter": "^10.4.0",
     "@carbon/scss-generator": "^10.11.0",
