+++ packages/elements/package.json
@@ -31,7 +31,7 @@
     "access": "public"
   },
   "scripts": {
-    "build": "yarn clean && bundler bundle src/index.js --name CarbonElements && node tasks/build.js && carbon-cli check \"scss/*.scss\" && carbon-cli sassdoc \"scss/**/*.scss\" -i \"**/{_inlined,vendor}/**\"",
+    "build": "yarn clean && carbon-cli bundle src/index.js --name CarbonElements && node tasks/build.js && carbon-cli check \"scss/*.scss\" && carbon-cli sassdoc \"scss/**/*.scss\" -i \"**/{_inlined,vendor}/**\"",
     "clean": "rimraf es lib umd && node tasks/clean.js"
   },
   "dependencies": {
@@ -45,7 +45,6 @@
     "@carbon/type": "^10.19.0"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli": "^10.17.0",
     "fs-extra": "^8.1.0",
     "klaw-sync": "^6.0.0",
