+++ packages/grid/package.json
@@ -35,7 +35,6 @@
     "@carbon/layout": "^10.16.0"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli": "^10.17.0",
     "rimraf": "^3.0.0"
   },
