+++ packages/layout/package.json
@@ -25,11 +25,10 @@
     "access": "public"
   },
   "scripts": {
-    "build": "yarn clean && bundler bundle src/index.js --name CarbonLayout && node tasks/build.js && carbon-cli sassdoc \"scss/**/*.scss\"",
+    "build": "yarn clean && carbon-cli bundle src/index.js --name CarbonLayout && node tasks/build.js && carbon-cli sassdoc \"scss/**/*.scss\"",
     "clean": "rimraf es lib umd scss/generated"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli-reporter": "^10.4.0",
     "@carbon/scss-generator": "^10.11.0",
     "@carbon/test-utils": "^10.13.0",
