+++ packages/motion/package.json
@@ -25,11 +25,10 @@
     "access": "public"
   },
   "scripts": {
-    "build": "yarn clean && bundler bundle src/index.js --name CarbonMotion && carbon-cli sassdoc \"scss/*.scss\"",
+    "build": "yarn clean && carbon-cli bundle src/index.js --name CarbonMotion && carbon-cli sassdoc \"scss/*.scss\"",
     "clean": "rimraf es lib umd"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli": "^10.17.0",
     "rimraf": "^3.0.0"
   },
