+++ packages/type/package.json
@@ -32,7 +32,7 @@
     "access": "public"
   },
   "scripts": {
-    "build": "yarn clean && bundler bundle src/index.js --name CarbonType && carbon-cli sassdoc \"scss/**/*.scss\" -i \"**/{_inlined,vendor}**\" && carbon-cli inline && carbon-cli check \"scss/*.scss\"",
+    "build": "yarn clean && carbon-cli bundle src/index.js --name CarbonType && carbon-cli sassdoc \"scss/**/*.scss\" -i \"**/{_inlined,vendor}**\" && carbon-cli inline && carbon-cli check \"scss/*.scss\"",
     "clean": "rimraf css es lib umd scss/_inlined scss/vendor"
   },
   "dependencies": {
@@ -40,7 +40,6 @@
     "@carbon/layout": "^10.16.0"
   },
   "devDependencies": {
-    "@carbon/bundler": "^10.11.0",
     "@carbon/cli": "^10.17.0",
     "@carbon/test-utils": "^10.13.0",
     "change-case": "^4.1.1",
